+++ Libplanet.Explorer.Executable/Program.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Bencodex.Types;
 using Cocona;
 using Libplanet.Action;
+using Libplanet.Assets;
 using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Blocks;
@@ -396,6 +397,9 @@ If omitted (default) explorer only the local blockchain store.")]
             }
 
             public IAction BlockAction => _impl.BlockAction;
+
+            /// <inheritdoc cref="IBlockPolicy{T}.NativeTokens"/>
+            public IImmutableSet<Currency> NativeTokens => ImmutableHashSet<Currency>.Empty;
 
             public IComparer<IBlockExcerpt> CanonicalChainComparer =>
                 _impl.CanonicalChainComparer;